renamed to 'componenets' to 'components'
