+++ frontend/app/components/GenerateKeyModal.tsx
